[UPDATED] added my camera object

+++ api/controllers/things/destroy-one-thing.js
@@ -17,11 +17,23 @@ module.exports = {
 
 
   exits: {
-
+    forbidden: {
+      description: 'The user making this request doesn\'t have the permissions to delete this things.',
+      responseType: 'forbidden'
+    }
   },
 
 
   fn: async function (inputs, exits) {
+
+    let thing = await Thing.findOne({
+      id: inputs.id
+    });
+
+    if(thing.owner !== this.req.me.id) {
+      throw 'forbidden';
+    }
+
     await Thing.destroy({id: inputs.id});
     // All done.
     return exits.success();
